Move execution result types from runner-executor-core to worker-core
