+++ crates/lib/backend-postgres/Cargo.toml
@@ -19,7 +19,6 @@ waymark-observability = { workspace = true }
 waymark-proto = { workspace = true }
 waymark-runner = { workspace = true }
 waymark-runner-execution-core = { workspace = true }
-waymark-runner-executor-core = { workspace = true }
 waymark-runner-expr-fmt = { workspace = true }
 waymark-runner-state = { workspace = true }
 waymark-scheduler-backend = { workspace = true }
@@ -33,6 +32,7 @@ waymark-vm-runtime-effect = { workspace = true }
 waymark-vm-runtime-promise-core = { workspace = true }
 waymark-webapp-backend = { workspace = true }
 waymark-webapp-core = { workspace = true }
+waymark-worker-core = { workspace = true }
 waymark-worker-status-backend = { workspace = true }
 waymark-workflow-completion-backend = { workspace = true }
 waymark-workflow-registry-backend = { workspace = true }
