+++ crates/lib/runner-executor-core/Cargo.toml
@@ -5,5 +5,4 @@ version.workspace = true
 publish.workspace = true
 
 [dependencies]
-serde = { workspace = true, features = ["derive"] }
-serde_json = { workspace = true }
+waymark-worker-core = { workspace = true }
