+++ crates/lib/worker-inline/Cargo.toml
@@ -6,7 +6,6 @@ publish.workspace = true
 
 [dependencies]
 waymark-observability = { workspace = true }
-waymark-runner-executor-core = { workspace = true }
 waymark-worker-core = { workspace = true }
 
 nonempty-collections = { workspace = true }
