+++ crates/lib/worker-inline/src/lib.rs
@@ -8,7 +8,7 @@ use serde_json::Value;
 use tokio::sync::mpsc;
 
 use waymark_observability::obs;
-use waymark_runner_executor_core::UncheckedExecutionResult;
+use waymark_worker_core::UncheckedExecutionResult;
 use waymark_worker_core::{
     ActionCompletion, ActionRequest, BaseWorkerPool, WorkerPoolError, error_to_value,
 };
